requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
et-xmlfile==1.0.1
html5lib==1.1
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
openpyxl==3.0.4
Pillow==7.2.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
six==1.15.0
tinycss2==1.0.2
Wand==0.6.1
WeasyPrint==51
webencodings==0.5.1
